+++ qx-api/src/main/java/org/qubic/qx/api/controller/AssetsController.java
@@ -2,9 +2,12 @@ package org.qubic.qx.api.controller;
 
 import org.qubic.qx.api.controller.service.AssetsService;
 import org.qubic.qx.api.db.domain.Asset;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static org.qubic.qx.api.redis.QxCacheManager.CACHE_NAME_ASSETS;
 
 @CrossOrigin
 @RestController
@@ -17,13 +20,10 @@ public class AssetsController {
         this.assetsService = assetsService;
     }
 
+    @Cacheable(CACHE_NAME_ASSETS)
     @GetMapping("/assets")
-    public List<Asset> getAssets(@RequestParam(name = "all", defaultValue = "false") boolean all) {
-        if (all) {
+    public List<Asset> getAssets() {
             return assetsService.getAllAssets();
-        } else {
-            return assetsService.getVerifiedAssets();
-        }
     }
 
 }