+++ qx-api/src/main/java/org/qubic/qx/api/db/domain/Asset.java
@@ -25,6 +25,7 @@ public class Asset implements Serializable {
 
     private String name;
 
+    @JsonIgnore
     private boolean verified;
 
 }