+++ qx-api/src/main/resources/db/migration/V15__delete_qpool_asset.sql
@@ -1,0 +1,13 @@
+-- doesn't delete transactions
+
+delete
+from trades
+where asset_id = (select id
+                  from assets
+                  where issuer = 'AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAFXIB'
+                    and name = 'QPOOL');
+
+delete
+from assets
+where issuer = 'AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAFXIB'
+  and name = 'QPOOL';